+++ modules/abstract-lightning/src/lightning/index.ts
@@ -2,3 +2,4 @@ export * from './signableJson';
 export * from './signature';
 export * from './lightningUtils';
 export * from './codecs';
+export * from './wpCodecs';