+++ modules/express/src/lightning/codecs.ts
@@ -36,11 +36,9 @@ export type GetWalletStateResponse = t.TypeOf<typeof GetWalletStateResponse>;
 export const InitLightningWalletRequest = t.intersection(
   [
     t.strict({
-      walletId: t.string,
       passphrase: t.string,
       signerIp: IPAddress,
       signerTlsCert: t.string,
-      expressIp: IPAddress,
     }),
     t.partial({
       signerTlsKey: t.string,
@@ -54,7 +52,6 @@ export type InitLightningWalletRequest = t.TypeOf<typeof InitLightningWalletRequ
 
 export const CreateSignerMacaroonRequest = t.strict(
   {
-    walletId: t.string,
     passphrase: t.string,
     watchOnlyIp: IPAddress,
   },
@@ -83,7 +80,6 @@ export type BakeMacaroonResponse = t.TypeOf<typeof BakeMacaroonResponse>;
 
 export const UnlockLightningWalletRequest = t.strict(
   {
-    walletId: t.string,
     passphrase: t.string,
   },
   'UnlockLightningWalletRequest'