+++ modules/express/src/lightning/lightningSignerRoutes.ts
@@ -19,6 +19,7 @@ import {
   UnlockLightningWalletRequest,
 } from './codecs';
 import { LndSignerClient } from './lndSignerClient';
+import { ApiResponseError } from '../errors';
 
 type Decrypt = (params: { input: string; password: string }) => string;
 
@@ -38,7 +39,10 @@ function getSignerRootKey(
   network: utxolib.Network,
   decrypt: Decrypt
 ) {
-  const userMainnetPrv = decrypt({ password: passphrase, input: userMainnetEncryptedPrv });
+  const userMainnetPrv = decrypt({
+    password: passphrase,
+    input: userMainnetEncryptedPrv,
+  });
   return utxolib.bitgo.keyutil.convertExtendedKeyNetwork(userMainnetPrv, utxolib.networks.bitcoin, network);
 }
 
@@ -54,25 +58,29 @@ function getMacaroonRootKey(passphrase: string, nodeAuthEncryptedPrv: string, de
  * Handle the request to initialise remote signer LND for a wallet.
  */
 export async function handleInitLightningWallet(req: express.Request): Promise<unknown> {
-  const { walletId, passphrase, signerTlsKey, signerTlsCert, signerIp, expressIp } = decodeOrElse(
+  const bitgo = req.bitgo;
+  const coinName = req.params.coin;
+  if (!isLightningCoinName(coinName)) {
+    throw new ApiResponseError(`Invalid coin ${coinName}. This is not a lightning coin.`, 400);
+  }
+  const coin = bitgo.coin(coinName);
+
+  const walletId = req.params.id;
+  if (typeof walletId !== 'string') {
+    throw new ApiResponseError(`Invalid wallet id: ${walletId}`, 400);
+  }
+
+  const { passphrase, signerTlsKey, signerTlsCert, signerIp, expressIp } = decodeOrElse(
     InitLightningWalletRequest.name,
     InitLightningWalletRequest,
     req.body,
     (_) => {
       // DON'T throw errors from decodeOrElse. It could leak sensitive information.
-      throw new Error('Invalid request body to initialise lightning wallet');
+      throw new ApiResponseError('Invalid request body to initialize lightning wallet', 400);
     }
   );
 
   const lndSignerClient = await LndSignerClient.create(walletId, req.config);
-
-  const bitgo = req.bitgo;
-  const coinName = req.params.coin;
-  if (!isLightningCoinName(coinName)) {
-    throw new Error(`Invalid coin to initialise lightning wallet: ${coinName}`);
-  }
-  const coin = bitgo.coin(coinName);
-
   const lightningWallet = getLightningWallet(await coin.wallets().get({ id: walletId }));
 
   const userKey = await lightningWallet.getLightningKeychain();
@@ -83,7 +91,9 @@ export async function handleInitLightningWallet(req: express.Request): Promise<u
   const macaroonRootKey = getMacaroonRootKey(passphrase, nodeAuthKey.encryptedPrv, bitgo.decrypt);
 
   const { admin_macaroon: adminMacaroon } = await lndSignerClient.initWallet({
-    wallet_password: passphrase,
+    // The passphrase at LND can only accommodate a base64 character set
+    // For more information, see BTC-1851
+    wallet_password: Buffer.from(passphrase).toString('base64'),
     extended_master_key: signerRootKey,
     macaroon_root_key: macaroonRootKey,
   });
@@ -111,33 +121,40 @@ export async function handleInitLightningWallet(req: express.Request): Promise<u
  * Handle the request to create a signer macaroon from remote signer LND for a wallet.
  */
 export async function handleCreateSignerMacaroon(req: express.Request): Promise<unknown> {
-  const { walletId, passphrase, watchOnlyIp } = decodeOrElse(
+  const bitgo = req.bitgo;
+  const coinName = req.params.coin;
+  if (!isLightningCoinName(coinName)) {
+    throw new ApiResponseError(`Invalid coin to create signer macaroon: ${coinName}. Must be a lightning coin.`, 400);
+  }
+  const coin = bitgo.coin(coinName);
+  const walletId = req.params.id;
+  if (typeof walletId !== 'string') {
+    throw new ApiResponseError(`Invalid wallet id: ${walletId}`, 400);
+  }
+
+  const { passphrase, watchOnlyIp } = decodeOrElse(
     CreateSignerMacaroonRequest.name,
     CreateSignerMacaroonRequest,
     req.body,
     (_) => {
       // DON'T throw errors from decodeOrElse. It could leak sensitive information.
-      throw new Error('Invalid request body to create signer macaroon');
+      throw new ApiResponseError('Invalid request body to create signer macaroon', 400);
     }
   );
 
   const lndSignerClient = await LndSignerClient.create(walletId, req.config);
-
-  const bitgo = req.bitgo;
-  const coinName = req.params.coin;
-  if (!isLightningCoinName(coinName)) {
-    throw new Error(`Invalid coin to create signer macaroon: ${coinName}`);
-  }
-  const coin = bitgo.coin(coinName);
 
   const wallet = await coin.wallets().get({ id: walletId });
   const lightningWallet = getLightningWallet(wallet);
 
   const encryptedSignerAdminMacaroon = wallet.coinSpecific()?.encryptedSignerAdminMacaroon;
   if (!encryptedSignerAdminMacaroon) {
-    throw new Error('Missing encryptedSignerAdminMacaroon in wallet');
+    throw new ApiResponseError('Missing encryptedSignerAdminMacaroon in wallet', 400);
   }
-  const adminMacaroon = bitgo.decrypt({ password: passphrase, input: encryptedSignerAdminMacaroon });
+  const adminMacaroon = bitgo.decrypt({
+    password: passphrase,
+    input: encryptedSignerAdminMacaroon,
+  });
 
   const { userAuthKey } = await lightningWallet.getLightningAuthKeychains();
 
@@ -147,7 +164,10 @@ export async function handleCreateSignerMacaroon(req: express.Request): Promise<
     lndSignerClient
   );
 
-  const userAuthXprv = bitgo.decrypt({ password: passphrase, input: userAuthKey.encryptedPrv });
+  const userAuthXprv = bitgo.decrypt({
+    password: passphrase,
+    input: userAuthKey.encryptedPrv,
+  });
 
   const encryptedSignerMacaroon = bitgo.encrypt({
     password: deriveLightningServiceSharedSecret(coinName, userAuthXprv).toString('hex'),
@@ -167,11 +187,11 @@ export async function handleCreateSignerMacaroon(req: express.Request): Promise<
 export async function handleGetLightningWalletState(req: express.Request): Promise<GetWalletStateResponse> {
   const coinName = req.params.coin;
   if (!isLightningCoinName(coinName)) {
-    throw new Error(`Invalid coin to get lightning wallet state: ${coinName}`);
+    throw new ApiResponseError(`Invalid coin to get lightning wallet state: ${coinName}`, 400);
   }
   const walletId = req.params.id;
   if (typeof walletId !== 'string') {
-    throw new Error(`Invalid wallet id: ${walletId}`);
+    throw new ApiResponseError(`Invalid wallet id: ${walletId}`, 400);
   }
 
   const lndSignerClient = await LndSignerClient.create(walletId, req.config);
@@ -182,21 +202,29 @@ export async function handleGetLightningWalletState(req: express.Request): Promi
  * Handle the request to unlock a wallet in the signer.
  */
 export async function handleUnlockLightningWallet(req: express.Request): Promise<void> {
-  const { walletId, passphrase } = decodeOrElse(
+  const coinName = req.params.coin;
+  if (!isLightningCoinName(coinName)) {
+    throw new ApiResponseError(`Invalid coin to unlock lightning wallet: ${coinName}`, 400);
+  }
+  const walletId = req.params.id;
+  if (typeof walletId !== 'string') {
+    throw new ApiResponseError(`Invalid wallet id: ${walletId}`, 400);
+  }
+
+  const { passphrase } = decodeOrElse(
     UnlockLightningWalletRequest.name,
     UnlockLightningWalletRequest,
     req.body,
     (_) => {
       // DON'T throw errors from decodeOrElse. It could leak sensitive information.
-      throw new Error('Invalid request body to unlock lightning wallet');
+      throw new ApiResponseError('Invalid request body to unlock lightning wallet', 400);
     }
   );
 
-  const coinName = req.params.coin;
-  if (!isLightningCoinName(coinName)) {
-    throw new Error(`Invalid coin to unlock lightning wallet: ${coinName}`);
-  }
-
   const lndSignerClient = await LndSignerClient.create(walletId, req.config);
-  return await lndSignerClient.unlockWallet({ wallet_password: passphrase });
+  // The passphrase at LND can only accommodate a base64 character set
+  // For more information, see BTC-1851
+  return await lndSignerClient.unlockWallet({
+    wallet_password: Buffer.from(passphrase).toString('base64'),
+  });
 }