+++ modules/sdk-coin-zeta/test/unit/zeta.ts
@@ -1,8 +1,8 @@
 import { CosmosTransaction, RedelegateMessage, SendMessage } from '@bitgo/abstract-cosmos';
 import { BitGoAPI } from '@bitgo/sdk-api';
+import { Environments } from '@bitgo/sdk-core';
 import { EcdsaRangeProof, EcdsaTypes } from '@bitgo/sdk-lib-mpc';
 import { TestBitGo, TestBitGoAPI, mockSerializedChallengeWithProofs } from '@bitgo/sdk-test';
-import { Environments } from '@bitgo/sdk-core';
 import { coins } from '@bitgo/statics';
 import BigNumber from 'bignumber.js';
 import { beforeEach } from 'mocha';
@@ -62,12 +62,6 @@ describe('Zeta', function () {
       addressDetails.memoId.should.equal('2');
     });
 
-    it('should throw on invalid memo id address', () => {
-      (() => {
-        basecoin.getAddressDetails(address.invalidMemoIdAddress);
-      }).should.throw();
-    });
-
     it('should throw on multiple memo id address', () => {
       (() => {
         basecoin.getAddressDetails(address.multipleMemoIdAddress);
@@ -95,7 +89,6 @@ describe('Zeta', function () {
       should.equal(utils.isValidAddress(undefined as unknown as string), false);
       should.equal(utils.isValidAddress(''), false);
       should.equal(utils.isValidAddress(address.validMemoIdAddress), true);
-      should.equal(utils.isValidAddress(address.invalidMemoIdAddress), false);
       should.equal(utils.isValidAddress(address.multipleMemoIdAddress), false);
     });