+++ modules/abstract-substrate/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.6](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-substrate@1.1.5...@bitgo/abstract-substrate@1.1.6) (2025-02-20)
+
+**Note:** Version bump only for package @bitgo/abstract-substrate
+
 ## [1.1.5](https://github.com/BitGo/BitGoJS/compare/@bitgo/abstract-substrate@1.1.3...@bitgo/abstract-substrate@1.1.5) (2025-02-19)
 
 ### Bug Fixes