+++ modules/sdk-coin-algo/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.65](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-algo@2.1.64...@bitgo/sdk-coin-algo@2.1.65) (2025-02-20)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-algo
+
 ## [2.1.64](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-algo@2.1.63...@bitgo/sdk-coin-algo@2.1.64) (2025-02-19)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-algo